ipa: rkisp1: awb: Store color temperature as an integer
The color temperature doesn't need floating point precision, and is
calculated by Awb::estimateCCT() as an unsigned integer. Store it with
the same data type in the frame context.

+++ src/ipa/rkisp1/ipa_context.h
@@ -100,7 +100,7 @@ struct IPAFrameContext : public FrameContext {
 			double blue;
 		} gains;
 
-		double temperatureK;
+		unsigned int temperatureK;
 		bool autoEnabled;
 	} awb;
